restructure | move itemmanager into utils
