+++ src/main/java/module-info.java
@@ -15,8 +15,8 @@ module david.buzas.whattoeat {
     exports david.buzas.whattoeat;
     exports david.buzas.whattoeat.entities;
     opens david.buzas.whattoeat.entities to javafx.fxml;
-    exports david.buzas.whattoeat.itemmanagement;
-    opens david.buzas.whattoeat.itemmanagement to javafx.fxml;
+    exports david.buzas.whattoeat.utils.itemmanager;
+    opens david.buzas.whattoeat.utils.itemmanager to javafx.fxml;
     exports david.buzas.whattoeat.repositories;
     opens david.buzas.whattoeat.repositories to javafx.fxml;
     exports david.buzas.whattoeat.ui.controllers;
